Feat: Rendering Elements Conditionally
I'm working with React's approach to rendering elements using
conditions. Unlike Angular's 'ngIf', React's approach is to use ternary
operators with JSX. It is important to note that JSX does not support
block statements, so regular if conditions will not work. I will look
into another way to do this from the controller or pure JavaScript side
rather than relying on JSX because it's kind of messy.