Rendering Conditional Elements (Ver. 2)
Implemented a more eloquent way of rendering conditional elements by
encapsulating the elements in a variable and using a conditional inside
the render() method to generate the variable onto the DOM. This made my
code a lot cleaner too. It is also a more standard way of doing this,
so I like it.