Feat: Replaced Radium with CSS Modules
Radium was taken off the project and the same styling features were
implemented using CSS Modules. So far, CSS Modules seems to be a better
approach as it is cleaner and more organized with its scoping feature on
each component. Separating the styles into CSS instead of doing them
inline keeps the component code a lot cleaner.