Feat: This.props vs props
A neat feature with React container components is the built in
this.props that is the same as a functional component's props. It might
be because the functional components are arrow functions without a scope
of its own to use a 'this' reference with. Another thing to note is
that functional components are more prevalent because they perform an
action and can't access state directly. This isolates containers
components as much as possible since they are the ones managing and
accessing the state.