Other Features
Some useful features for improving code and app experience are prop
types and references. Prop types are a good way to ensure that props are
hard-typed when they are passed, returning errors if a certain prop
is a different type when it was passed in. References or 'ref' are
mainly used to reference an element directly and pass it to a variable
(within a stateful class component only). Refs are useful for doing
small things such as focusing on an input, but should not be used for
styling and other element-specific tasks that can be handled more
efficiently.